Add fixed top bar and reset team data
