+++ Ballog/Views/FeedView.swift
@@ -3,7 +3,10 @@ import SwiftUI
 /// 피드 탭은 현재 준비 중이다.
 struct FeedView: View {
     var body: some View {
-        Text("피드 준비 중")
+        NavigationStack {
+            Text("피드 준비 중")
+        }
+        .ballogTopBar()
     }
 }
