+++ Ballog/Views/MainHomeView.swift
@@ -52,12 +52,6 @@ struct MainHomeView: View {
         return try? JSONDecoder().decode(ProfileCard.self, from: data)
     }
 
-    private var todayString: String {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "ko_KR")
-        formatter.dateFormat = "yyyy M월 d일 EEEE"
-        return formatter.string(from: Date())
-    }
 
     private var competitionEvents: [TeamEvent] {
         eventStore.events.filter { $0.type == .match }.sorted { $0.date < $1.date }
@@ -90,9 +84,6 @@ struct MainHomeView: View {
     var body: some View {
         NavigationStack {
             VStack(spacing: Layout.spacing) {
-                Spacer(minLength: 60) // 위 여백
-
-                topBar
                 quoteSection
                 scheduleSection
                 thisWeekScheduleSection // 추가된 부분
@@ -107,39 +98,9 @@ struct MainHomeView: View {
             .ignoresSafeArea()
             .onAppear { quote = quotes.randomElement() ?? "" }
         }
+        .ballogTopBar()
     }
 
-    private var topBar: some View {
-        HStack(spacing: 16) {
-            VStack(alignment: .leading, spacing: 2) {
-                Text("볼로그")
-                    .font(.title2.bold())
-                Text(todayString)
-                    .font(.caption)
-                    .foregroundColor(.secondary)
-            }
-            .padding(.leading)
-
-            Spacer()
-
-            NavigationLink(destination: ProfileView()) {
-                Image(systemName: "person.circle.fill")
-                    .resizable()
-                    .frame(width: 28, height: 28)
-            }
-            NavigationLink(destination: NotificationView()) {
-                Image(systemName: "bell")
-                    .resizable()
-                    .frame(width: 24, height: 24)
-            }
-            NavigationLink(destination: SettingsView()) {
-                Image(systemName: "gearshape")
-                    .resizable()
-                    .frame(width: 24, height: 24)
-            }
-        }
-        .padding(.vertical, 8)
-    }
 
     private var quoteSection: some View {
         VStack(spacing: 8) {
