+++ Ballog/Views/NotificationView.swift
@@ -13,7 +13,6 @@ struct NotificationView: View {
 
     var body: some View {
         NavigationStack {
-            Spacer(minLength: 100) // 위 여백
             List(posts, id: \.self) { post in
                 Text(post)
             }
@@ -24,6 +23,7 @@ struct NotificationView: View {
         }
         .background(Color.pageBackground)
         .ignoresSafeArea()
+        .ballogTopBar()
     }
 }
